backend/requirements.txt
django-5.0.7 
django-cors-headers-4.4.0
plaid-0.1.7
plaid_python==24.0.0